App/requirements.txt
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
jsonschema==4.17.3
email-validator==1.3.1
python-dotenv==1.0.0
pymysql==1.0.2
